De-dup permutation helpers
